Hacky unit/tuple variant fix
We sneak the type of the label into the ident itself; mnemonic being '('
for an "open" value, ')' for "closed".  The advantage to these over
something more meaningful (e.g. "U" and "T" or whatever) is that these
can't be parsed as part of an ident, and if we fail to remove them
somewhere in generated code, a syntax error will occur.